ANIME RECOMMENDER/requirements.txt
langchain
langchain-classic
langchain-community
langchain-text-splitters
langchain_groq
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain-huggingface
